+++ base/terminfo.jl
@@ -138,7 +138,7 @@ function extendedterminfo(data::IO, NumInt::Union{Type{Int16}, Type{Int32}})
     end
     labels = map(Symbol, _terminfo_read_strings(table_data, table_indices[string_count+1:end]))
     Dict{Symbol, Union{Bool, Int, String, Nothing}}(
-        zip(labels, Iterators.flatten((flags, numbers, strings))))
+        zip(labels, Iterators.cat(flags, numbers, strings)))
 end
 
 """